# perf: reduce cloning operations