--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,5 +18,10 @@ serde = "1.0.126"
 
 [dev-dependencies]
 pretty_assertions = "0.6.1"
-serde_derive = "1.0.115"
-serde_json = "1.0.57"
+serde_derive      = "1.0.115"
+serde_json        = "1.0.57"
+criterion         = {  version = "0.5.1", features = ["html_reports"] }
+
+[[bench]]
+name = "parser"
+harness = false