### Benchmarks: 

#### Before:
```
range                   time:   [633.57 ns 634.84 ns 636.11 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```

#### After:
```
range                   time:   [614.54 ns 615.52 ns 616.51 ns]
                        change: [-4.0570% -2.8459% -1.6207%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
```